requirements.txt
streamlit
openai
openai-whisper
python-dotenv
torch
torchaudio